Refactored the farmosUtil.js library  into separate smaller libraries  along with a core library
